backend/requirements.txt
Flask
Streamlit
PyMuPDF
pdfplumber
python-docx
docx2txt
spaCy
nltk
langchain
chromadb
openai
rank_bm25
fuzzywuzzy
langchain-openai
langchain-community
scikit-learn
transformers
sentence-transformers
